requirements.txt
fastapi
uvicorn
pydantic
python-multipart
websockets
requests
pillow
numpy
opencv-python
pywin32
pytesseract
python-dotenv
pywebview
